Refactor data management of imu plugin
