+++ plugins/imu/Imu.cc
@@ -1,24 +1,37 @@
 #include <ConfigurationHelpers.hh>
 #include <DeviceIdGenerator.hh>
 #include <ImuDriver.cpp>
+#include <ImuShared.hh>
+
+#include <memory>
+#include <mutex>
+#include <string>
 
 #include <gz/msgs/details/imu.pb.h>
 #include <gz/plugin/Register.hh>
+#include <gz/sim/Entity.hh>
+#include <gz/sim/EntityComponentManager.hh>
+#include <gz/sim/EventManager.hh>
 #include <gz/sim/Link.hh>
 #include <gz/sim/Model.hh>
 #include <gz/sim/Sensor.hh>
 #include <gz/sim/System.hh>
+#include <gz/sim/Types.hh>
 #include <gz/sim/Util.hh>
 #include <gz/sim/components/Imu.hh>
 #include <gz/sim/components/Name.hh>
 #include <gz/sim/components/ParentEntity.hh>
 #include <gz/sim/components/Sensor.hh>
 #include <gz/transport/Node.hh>
+#include <sdf/Element.hh>
 
+#include <yarp/dev/Drivers.h>
 #include <yarp/dev/PolyDriver.h>
 #include <yarp/dev/PolyDriverList.h>
+#include <yarp/os/Log.h>
 #include <yarp/os/LogStream.h>
 #include <yarp/os/Network.h>
+#include <yarp/os/Property.h>
 
 using namespace gz;
 using namespace sim;
@@ -49,7 +62,6 @@ public:
 
         if (m_imuDriver.isValid())
             m_imuDriver.close();
-        ImuDataSingleton::getHandler()->removeSensor(sensorScopedName);
     }
 
     virtual void Configure(const Entity& _entity,
@@ -104,9 +116,6 @@ public:
 
         driver_properties.put(YarpIMUScopedName.c_str(), sensorScopedName.c_str());
 
-        // Insert the pointer in the singleton handler for retriving it in the yarp driver
-        ImuDataSingleton::getHandler()->setSensor(&(this->imuData));
-
         driver_properties.put("device", "gazebo_imu");
         driver_properties.put("sensor_name", sensorName);
         if (!m_imuDriver.open(driver_properties))
@@ -114,6 +123,16 @@ public:
             yError() << "gz-sim-yarp-imu-system Plugin failed: error in opening yarp driver";
             return;
         }
+
+        IImuData* iImuDataPtr = nullptr;
+        auto viewOk = m_imuDriver.view(iImuDataPtr);
+
+        if (!viewOk || !iImuDataPtr)
+        {
+            yError() << "gz-sim-yarp-imu-system: failed to get IImuData interface";
+            return;
+        }
+        iImuDataPtr->setImuData(&imuData);
 
         auto deviceName = driver_properties.find("yarpDeviceName").asString();
         m_deviceId = DeviceIdGenerator::generateDeviceId(sensor, _ecm, deviceName);
@@ -124,6 +143,7 @@ public:
                      << ")";
             return;
         }
+
         m_deviceRegistered = true;
         yInfo() << "Registered YARP device with instance name:" << m_deviceId;
     }
@@ -176,7 +196,7 @@ private:
     std::string m_deviceId;
     std::string sensorScopedName;
     bool m_deviceRegistered;
-    ImuData imuData;
+    ::gzyarp::ImuData imuData;
     bool imuInitialized;
     gz::transport::Node node;
     gz::msgs::IMU imuMsg;
